test: add test iterator in base consume
